+++ src/tests/integration/test_mssql.py
@@ -113,19 +113,3 @@ def test_sqlplot_boxplot(ip_with_MSSQL, cell):
     out = ip_with_MSSQL.run_cell(cell)
 
     assert type(out.result).__name__ in {"Axes", "AxesSubplot"}
-
-
-def test_unsupported_function(ip_with_MSSQL, test_table_name_dict):
-    # clean current Axes
-    plt.cla()
-
-    with pytest.raises(UsageError) as excinfo:
-        ip_with_MSSQL.run_cell(
-            f"%sqlplot boxplot --table " f"{test_table_name_dict['taxi']} --column x"
-        )
-
-    assert "Ensure that percentile_disc function is available" in str(excinfo.value)
-    assert (
-        "If you need help solving this issue, "
-        "send us a message: https://ploomber.io/community" in str(excinfo.value)
-    )