+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.9.1dev
 * [Fix] Fix boxplot for duckdb native ([#728](https://github.com/ploomber/jupysql/issues/728))
+* [Feature] Adds Redshift support for `%sqlplot boxplot`
 
 ## 0.9.0 (2023-08-01)